+++ abc/abc151-200/src/abc198/abc198D.cpp
@@ -1,6 +1,5 @@
 #include <bits/stdc++.h>
 using namespace std;
-
 using ll = long long;
 
 const int N = 26;
@@ -22,9 +21,8 @@ ll process(string &s, vector<int> &dict) {
 	return res;
 }
 
-bool solve(string &s1, string &s2, string &s3, vector<char> &vars, vector<int> &dict, vector<bool> &used,
-		long unsigned int idx) {
-	if (idx == vars.size()) {
+bool solve(string &s1, string &s2, string &s3, vector<char> &vars, vector<int> &dict, vector<bool> &used, int idx) {
+	if (idx == int(vars.size())) {
 		return process(s1, dict) + process(s2, dict) == process(s3, dict);
 	}
 	char c = vars[idx];
