+++ abc/abc151-200/src/abc198/abc198F.cpp
@@ -2,25 +2,24 @@
 #include <atcoder/modint>
 #include <atcoder/convolution>
 using namespace std;
-
 using ll = long long;
 using mint = atcoder::modint998244353;
 
 mint bostan_mori(vector<mint> &p, vector<mint> &q, ll n) {
 	while (n > 0) {
 		vector<mint> inv_q(q.size());
-		for (long unsigned int i = 0; i < q.size(); i++) {
+		for (int i = 0; i < int(q.size()); i++) {
 			inv_q[i] = (0 == (i & 1)) ? q[i] : -q[i];
 		}
 		p = atcoder::convolution(p, inv_q);
 		q = atcoder::convolution(q, inv_q);
 		int j = 0;
-		for (long unsigned int i = (n & 1); i < p.size(); i += 2) {
+		for (int i = (n & 1); i < int(p.size()); i += 2) {
 			p[j++] = p[i];
 		}
 		p.resize(j);
 		j = 0;
-		for (long unsigned int i = 0; i < q.size(); i += 2) {
+		for (int i = 0; i < int(q.size()); i += 2) {
 			q[j++] = q[i];
 		}
 		q.resize(j);
